Move files to src/connect.
