+++ src/connect/parseFilter.js
